% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{normalmixEM}
\alias{is.wholenumber}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DistributionUtils}{\code{\link[DistributionUtils]{is.wholenumber}}}

  \item{mixtools}{\code{\link[mixtools]{normalmixEM}}}
}}

