% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztpln.r
\docType{package}
\name{ztpln-package}
\alias{ztpln}
\alias{ztpln-package}
\title{The zero-truncated compund poisson-lognormal distributions and their
mixtures}
\description{
Functions for obtaining the density, random deviates and maximum likelihood
estimates of the zero-truncated Poisson lognormal distributions and their
mixtures.
}
\references{
Bulmer, M. G. 1974. On Fitting the Poisson Lognormal Distribution to Species-Abundance Data. Biometrics 30:101-110.

Inouye, D., E. Yang, G. Allen, and P. Ravikumar. 2017. A Review of Multivariate Distributions for Count Data Derived from the Poisson Distribution. Wiley interdisciplinary reviews. Computational statistics 9.

Raqab, M. Z., D. Kundu, and F. A. Al-Awadhi. 2019. Compound zero-truncated Poisson normal distribution and its applications. Communications in Statistics - Theory and Methods:1–21.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattocci27/ztpln}
  \item Report bugs at \url{https://github.com/mattocci27/ztpln/issues}
}

}
\author{
Masatoshi Katabuchi \href{mailto:mattocci27@gmail.com}{mattocci27@gmail.com}
}
\keyword{internal}
