% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztpln.r
\docType{package}
\name{ztpln-package}
\alias{ztpln}
\alias{ztpln-package}
\title{The zero-truncated compund poisson-lognormal distributions and their
mixtures}
\description{
Functions for obtaining the density, random deviates and maximum likelihood
estimates of the zero-truncated Poisson lognormal distributions and their
mixtures.
}
\references{
M. G. Bulmer. 1974. On Fitting the Poisson Lognormal Distribution to Species-Abundance Data. Biometrics 30: 101-110 \href{https://www.jstor.org/stable/2529621?origin=crossref&seq=1#metadata_info_tab_contents}{doi:10.2307/2529621}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattocci27/ztpln}
  \item Report bugs at \url{https://github.com/mattocci27/ztpln/issues}
}

}
\author{
Masatoshi Katabuchi \href{mailto:mattocci27@gmail.com}{mattocci27@gmail.com}
}
\keyword{internal}
